refactor: split run_test.go into focused test files
Split 1914-line run_test.go into:
- helpers_test.go (shared test utilities)
- crd_sync_test.go
- kubernetes_scraper_test.go (dedup, relationships, fixtures, full)
- plugins_test.go
- external_entities_test.go (users, roles, groups, stale)
- config_access_test.go (user/role/group aliases, logs, RBAC)
